First attempt at blurriness calculation.
Works decently well. The current algorithm measures the blurriness in a given photo. First, it tries to detect a face. If a face is detected, then ONLY the areas of the face are considered for blurriness (the idea is to find out if the face is in focus, if there are multiple faces, the blurriness is averaged across all faces). Then, if no face is found, simply calculate the blurriness across the entire image. Test this using the following command:

python detect_blur.py